+++ tests/test_context.py
@@ -959,6 +959,7 @@ class TestOuterContextProperty:
         assert "outer_value" in rendered
 
 
+# TODO_v1: Remove, superseded by `component_vars.slots`
 @djc_test
 class TestContextVarsIsFilled:
     class IsFilledVarsComponent(Component):